Move locked_account fixture up in tree
